requirements.txt
Appium-Python-Client==2.11.1
langchain==0.0.267
openai==0.27.8
python-dotenv==1.0.0
uuid==1.30 